refactor : Events UI
일정 hooks 디렉토리로 일부 변경

#13
