+++ src/features/events/personal/components/modal/PersonalScheduleDetailsModal.css
